bugs fixed for post creation, added models and serializer info
