+++ engine/base_models/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BaseModelConfig(AppConfig):
+    name = 'base_model'
