+++ engine/post/models.py
@@ -5,7 +5,8 @@ from engine.account.models import User
 from engine.category.models import Category, Subcategory
 from engine.hashtag.models import Hashtag
 
-from engine.utils import BaseModel
+from engine.base_models.models import BaseModel
+
 
 class Post(BaseModel):
     name = models.CharField(max_length=254, blank=False)
@@ -13,12 +14,6 @@ class Post(BaseModel):
     category = models.ForeignKey(Category, null=True, editable=False, on_delete=models.SET_NULL)
     subcategories = models.ManyToManyField(Subcategory)
     hashtags = models.ManyToManyField(Hashtag)
-    """
-    created_by = models.ForeignKey(User, null=True, editable=False, on_delete=models.SET_NULL)
-    created_at = models.DateTimeField(default=now, editable=False)
-    changed = models.DateTimeField(default=now, editable=True)
-    active = models.BooleanField(default=False)
-    """
     
     class Meta:
         ordering = ('created_at',)
