+++ engine/utils.py
@@ -1,4 +1,5 @@
 from django.db import models
+from django.utils.timezone import now
 
 from rest_framework_jwt.serializers import VerifyJSONWebTokenSerializer
 
@@ -32,11 +33,5 @@ def get_user_token(func):
         return obj_result
     return func_wrapper
 
-
-class BaseModel(models.Model):
-    created_by = models.ForeignKey(User, null=True, editable=False, on_delete=models.SET_NULL)
-    created_at = models.DateTimeField(default=now, editable=False)
-    changed = models.DateTimeField(default=now, editable=True)
-    active = models.BooleanField(default=False)
 
 #upload to digital ocean
